Massive transients implementations
Input file with a list of the subhalos that are classified as transients and the halos that host them. When building the trees and making connections subhalo-descendant, if there is a transient in the descendant halo and several constraints are fulfilled, then default subhalo-descendant connection is broken and a subhalo-transient connection has priority. These criteria can be summarized as: 1. transient more massive than descendant, 2. subhalo-descendant connection has a severe mass loss, 3. subhalo-transient connection is smooth enough in terms of the mass. Define parameters used for the massive transients treatment and for computing statistics. Set to true the default flags to apply mass swapping anf massive transient fixes