Massive transient implementations II
Final changes to implement the massive transient treatment. Plus after playing with the parameters involved in the transient fixes, we set their default values considering mass changes more or less smooth: the mass ratio to break a main progenitor-subhalo link should be <0.7 while the mass ratio to create a new main progenitor-transient link should be in the range 0.1-3.