Define massive transients inside the code
Massive transient catalogues are not required right noe. The massive transients are flagged inside the code as the subhalos that belong to main branches that are born with an unexpectedly high number of particles: 1. find the subhalos with no main progenitor, 2. find subhalos with no progenitor that belong to main branches, 3. they are defined as transients if the particle number at birth is above the considered threshold. There are 3 different threshold in the particle number at birth for main branches: a constant value of 200 particles, a constant value of 10 times the minimum particle number for a subhalo, and 3 times the standard deviation of the particle number at birth for main branches at each snapshot. Set as a default the 3 sigma method.